+++ e2e/tests/binds.rs
@@ -1,6 +1,6 @@
-use e2e::test_utils::sk;
-use e2e::test_utils::Keys::*;
-use e2e::test_utils::TmuxController;
+use e2e::sk;
+use e2e::Keys::*;
+use e2e::TmuxController;
 use std::io::Result;
 
 fn setup(input: &str, opts: &[&str]) -> Result<TmuxController> {
