+++ e2e/tests/tmux.rs
@@ -1,5 +1,5 @@
-use e2e::test_utils::Keys::*;
-use e2e::test_utils::TmuxController;
+use e2e::Keys::*;
+use e2e::TmuxController;
 use std::fs::File;
 use std::fs::Permissions;
 use std::io::Read;
